split voting validation logic
